.agent/skills/thesisflow-manager/requirements.txt
sentence-transformers
faiss-cpu
pypdf
langchain-text-splitters
numpy<1.27.0
paperscraper
biopython
arxiv
tenacity
rich
beautifulsoup4
